Everything is broken lol

+++ include/map.hpp
@@ -1,5 +1,5 @@
 #include <SFML/Graphics.hpp>
 
-sf::Vector2i size = {1920, 1080};
+sf::Vector2f size = {1920, 1080};
 int gridSize = 20;
 int gridLineThickness = 3;
